+++ lib/generators/cms/kickstart/templates/app/models/obj.rb
@@ -1,7 +1,7 @@
 require './lib/rails_connector/cms_attributes'
 
 # This class represents the base class of all CMS objects and implements behavior that all CMS
-# objects, regardless whether they are pages, boxes or resources have in common.
+# objects, regardless whether they are pages, widgets or resources have in common.
 class Obj < ::RailsConnector::BasicObj
   include RailsConnector::CmsAttributes